Fatemeh/requirements.txt
numpy
pandas
matplotlib
scikit-learn
torch
torchvision
timm
jupyterlab
notebook
ipykernel
pyyaml
tqdm
pillow